# http: add scry and noun caches