The runtime component of the Eyre cache.

This is a draft as I still have a few open questions (and haven't tested anything yet, except confirming that this at least compiles...). Questions are in-line below.