@lukechampine 2023-03-18
Updated and tested; basic functionality seems to be working!

Add a cached response:
```
> |pass [%e %set-response '/foo' `[%.n %payload [200 ~] `11+'foo-payload']]
```
Scry it:
```
> .^(cache-entry:eyre %ex /=//=/cache/1/~~~2f.foo)
[ auth=%.n
    body
  [ %payload
      simple-payload
    [ response-header=[status-code=200 headers=~]
      data=[~ [p=11 q=121.352.706.915.279.217.521.291.110]]
    ]
  ]
]
```
And curl it:
```
$ curl localhost:80/foo
foo-payload
```
Logging confirms that the response is being served from the noun cache, not from Eyre.


However, when the entry is a `%scry`, I get a strange crash:
```
> |pass [%e %set-response 'bail' `[%.n %scry /ex/~zod//~2023.3.18..04.30.04..8101/cors]
> .^(cache-entry:eyre %ex /=//=/cache/1/~~bail)
Assertion '0 != u3R->ski.gul' failed in pkg/noun\manage.c:1256
Assertion failed: (0 != u3a_Road->ski.gul), function u3m_soft_esc, file manage.c, line 1256.
```
Not sure what's going on here, will investigate further.