@lukechampine 2023-03-18
Oh yeah, let's try benchmarking it!

Without cache:
```
$ go-wrk -c 2 -d 5 http://localhost:80/foo
Running 5s test @ http://localhost:80/foo
  2 goroutine(s) running concurrently
1008 requests in 5.003845417s, 107.30KB read
Requests/sec:           201.45
Transfer/sec:           21.44KB
Avg Req Time:           9.928264ms
Fastest Request:        6.5115ms
Slowest Request:        75.220166ms
Number of Errors:       0
```

With cache:
```
$ go-wrk -c 2 -d 5 http://localhost:80/foo
Running 5s test @ http://localhost:80/foo
  2 goroutine(s) running concurrently
56713 requests in 4.976299745s, 5.90MB read
Requests/sec:           11396.62
Transfer/sec:           1.18MB
Avg Req Time:           175.49µs
Fastest Request:        123.125µs
Slowest Request:        5.91275ms
Number of Errors:       0
```
A 50x speedup!