@belisarius222 2023-03-19
50x!  Nice.  Great to see this is running.

The crash is coming from having a .^ in Eyre.  Kernel code is run without a scry handler, so a .^ will always crash in kernelspace.  I think I led you astray there, apologies.  I should have recommended (rof ...) instead of dotket.