@lukechampine 2023-03-22
ok, I fixed the scry handling in https://github.com/urbit/urbit/pull/5927, so this is fully-functional now!