@vcavallo 2023-04-12
https://github.com/urbit/vere/pull/259#issuecomment-1474883042

I'm salivating